week_13/day_1/creating_interfaces/code/interfaces_start_point/src/test/java/CarTest.java
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

public class CarTest {

    private Car car;

    @Before
    public void before() {
        car = new Car("Ford", "Escort");
    }

    @Test
    public void hasMake() {
        assertEquals("Ford", car.getMake());
    }

    @Test
    public void hasModel() {
        assertEquals("Escort", car.getModel());
    }

    @Test
    public void odomoterReadingStartsAtZero() {
        assertEquals(0, car.getOdometerReading());
    }

    @Test
    public void canMove() {
        car.move(10);
        assertEquals(10,car.getOdometerReading());
    }

    @Test
    public void canStart() {
        assertEquals("Vroom",car.start());
    }
}
